Compiles now, implemented assembly for syscall stuff

+++ student-distrib/fs.h
@@ -65,6 +65,6 @@ uint32_t read_directory(uint32_t offset, uint8_t* buf, uint32_t length);
 int32_t read_data(uint32_t inode, uint32_t offset, uint8_t* buf, uint32_t length);
 
 /* Loads an executable file into correct location in memory */
-int32_t load(uint8_t inode, uint32_t* addr);
+int32_t load(uint32_t inode, uint8_t* addr);
 
 #endif /* _FS_H */
